LayoutTests/media/video-load-preload-metadata-expected.txt


Test that preload="metadata" disables buffering before playback.

EXPECTED (video.preload == 'metadata') OK
EVENT(loadedmetadata)
EXPECTED (video.readyState < '4') OK
EXPECTED (video.buffered.length == '0') OK
RUN(video.play())
EVENT(canplaythrough)
TEST(playbackStarted) OK
EXPECTED (video.buffered.length > '0') OK
END OF TEST
